Fix a hanging issue when processing empty data.
The output iterator will wait on the batch queue when calling `hasNext`,
and suppose to be waked up when the Python runner inserts something into
the batch queue. But the insertion will never happen if the input data
is empty. So it hangs forever.

The solution is to let the Python runner always wake up the output iterator
after it finishes the data writing by calling the new added API `finish()`.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>