Build python output schema from udf expressions
Because the result attributes are NOT always equal to the python output schema.

For example, on databricks when projecting only one column from a python UDF output
where containing multiple result columns, there will be only one attribute in the
result attributes for the projecting output, but the output schema for this
python udf contains multiple columns.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>