WindowInPandas supports sending multiple groups
Let WindowInPandas support sending all the groups in a batch one time
to Python for better performance.

Thanks to the Python implementation, we can do this without Python change for
bounded windows, by having the window bounds been relative to the head of the
whole batch, instead of the head of its group.

Disable it by default due to some limitations.

- Add a config to enable/disable this feature.
- Move bounds building into iterator instead of the methods of the plan
  to avoid serializing the whole plan.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>