Switch to single-level structs in hash aggregate test
Closes #2428. Due to a subtle difference between the single-column
gen_df and two_col_df copy-and-pasting of the StructGen between tests
introduced testing of double-nestedd structs in single-column tests
using gen_df. This seems to trigger a bug in Spark with deeper-nested
columns that we can invesigate later. For now just we just restore the
intended nestingn degree 1 in the affected test.

Signed-off-by: Gera Shegalov <gera@apache.org>