skip test_window_aggs_for_rows_lead_lag_on_arrays
If some rows of order-by columns (it's `a,b,c` in the test) are equal,
then it may fail because CPU and GPU can't guarantee the order for the
same rows, while lead/lag is typically depending on row's order.

The solution is we should add the aggregation column `d` and the default column
`d_default` columns into the order-by to guarantee the order. But for now,
sorting on array has not been supported yet, see
https://github.com/NVIDIA/spark-rapids/issues/2470.

So this PR just skip the test

Signed-off-by: Bobby Wang <wbo4958@gmail.com>