Explicitly export vars for the Databricks test script
There is a mismatch of param numbers in the Databricks test script,
when running Databricks nightly test job or nightly build job.

The test job takes 1 to 3 params, while the build job requires only 1 to 2 params.
This makes it hard to match up params between the test job and the build jobs.

To fix the issue, we explicitly export vars for the test script instead of taking shell params.
This method also makes it easier to extend vars/params for the test script.

Signed-off-by: Tim Liu <timl@nvidia.com>