Fix childExprs list for GpuWindowExpression, for Spark 3.1.x.
Fixes #2939.

When the type-checking for expressions in `RapidsMeta` was changed in #2888 (i.e. commit 1a093229),
`OffsetWindowFunctionMeta` was modified to include `input`, `offset`, and `default` expressions among
its `childExprs`.
Unfortunately, the commit neglected to replicate that change in `Spark311Shims` `OffsetWindowFunctionMeta`.

The error manifests as a failure to do type-checks for `LEAD()`/`LAG()` window functions, with the error:
`java.lang.AssertionError: assertion failed: Lead expected at least 3 but found 0`.

This commit should remedy the situation, and allow `LEAD()`/`LAG()` to function correctly.

Signed-off-by: Mithun RK <mythrocks@gmail.com>