Fix column->row conversion GPU check: (#5190)
Fixes #5181.

The change in #5122 modifies the row->column conversion code to run
differently on Pascal GPUs. It does so by checking the GPU architecture
to detect whether the current device is Pascal.
In a distributed setup, this check should run on the Spark executor. As it
currently stands, the check is erroneously run on the driver.

This fix postpones the GPU architecture check till the executor task attempts
to fetch result rows.

Signed-off-by: MithunR <mythrocks@gmail.com>