Fix CPU fallback for custom timestamp formats in Hive text tables: (#7305)
Fixes #7303.

The timestamp parsing behaviour was altered in #7265 to support the default
timestamp formats (i.e. yyyy-MM-dd HH:mm:ss[.SSS]). The intended behaviour was to
fall back to CPU execution if the user specifies a custom (non-default)
timestamp format, via the `timestamp.formats` property.

A bug in #7265 causes only the SERDEPROPERTIES to be checked for the timestamp
property, and not TBLPROPERTIES. This error was masked by an error in the test
for the fallback.

This commit fixes both the fallback logic, and the error in its test.

Signed-off-by: MithunR <mythrocks@gmail.com>