Fix CPU fallback for GpuHiveTableScanExec:
Fixes #7374.

The Hive delimited text reader does not currently support nested types. When STRUCT, ARRAY, MAP, BINARY columns are encountered, the reader is expected fall back to CPU reads.

A bug in the fallback logic causes it to fail for tables with nested types, if the output projection excludes the nested type columns. This commit fixes the fallback logic.

Signed-off-by: MithunR <mythrocks@gmail.com>