Fix CPU fallback for GpuHiveTableScanExec (#7375)
* Fix CPU fallback for GpuHiveTableScanExec:

Fixes #7374.

The Hive delimited text reader does not currently support nested types. When STRUCT, ARRAY, MAP, BINARY columns are encountered, the reader is expected fall back to CPU reads.

A bug in the fallback logic causes it to fail for tables with nested types, if the output projection excludes the nested type columns. This commit fixes the fallback logic.

Signed-off-by: MithunR <mythrocks@gmail.com>

* Removed commented code.

* Review fixes:

1. Adjusted logic to flag unsupported types.
2. Improved variable names in tests.
3. Improved documentation of limitation of GpuHiveTableScanExec,
   in reading text files partially.

Signed-off-by: MithunR <mythrocks@gmail.com>