Allow Shims to override Hive expr/exec replacements.
If a distro shim attempts to provide a replacement for Hive execs, e.g.
`HiveTableScanExec`, it is likely to fail, because `GpuOverrides` currently
applies execs derived from the HiveProvider last.

This commit changes this order, so that the active shim has the last say
in exec replacements. This should allow for shims to selectively
replace or disable replacements for Hive operators.