Fix Parquet dynamic partition test for Databricks
Fixes #7679.
The Parquet dynamic partitioning test added in #7653 does not run on Databricks
successfully.
It turns out that on Databricks, a SQL query that writes partitions
dynamically to a Parquet table is routed through `InsertIntoHiveTable`
instead of `InsertIntoHadoopFSRelationCommand` as on Apache Spark.
For one, `InsertIntoHiveTable` is not currently supported on `spark-rapids`
(at least until #7556 is merged).
For another, `GpuInsertIntoHiveTable` would not support serialization of Parquet
tables anyway.

This commit skips the SQL dynamic partitioning test on Databricks, and replaces
it with a Dataframes-based test. The new test is functional on both Databricks
and Apache Spark.

Signed-off-by: MithunR <mythrocks@gmail.com>