Disable Dynamic partitioning tests on CDH:
Fixes #7693.

This commit disables two (failing) tests on CDH.

1. `test_partitioned_sql_parquet_write()` fails on CDH just like it fails on
Databricks. This is because SQL partition writes are routed through the
`InsertIntoHiveTable` instead of `InsertIntoHadoopFSRelation` as on
Apache Spark. `GpuInsertIntoHiveTable` only supports Hive text tables
at the moment. Further, the exec is disabled even for Hive text tables
on CDH. This test would be testing disabled functionality on CDH.

2. `test_partitioned_hive_text_write()` should have been disabled on CDH
as part of #7556. This was an oversight. CDH does not currently support
acceleration for reads or writes of Hive delimited text.

Signed-off-by: MithunR <mythrocks@gmail.com>