Workaround for incompatible serialization for `Inf` floats in Hive text write
Fixes #7686.

The `GpuHiveTextFileFormat` is currently serializing Float(Infinity) as 'Inf' instead of
'Infinity'. (It boils down to the CUDF implementation using 'Inf'.)

Apache Spark (and Apache Hive) expect Infinity strings to be represented as 'Infinity'.
'Inf' is read as `NULL`.

This commit adds special handling for floating point, such that 'Inf' is replaced with
'Infinity' before serializing to the table.

Signed-off-by: MithunR <mythrocks@gmail.com>