Set `spark.executor.cores` for integration tests.
Fixes #9135. (By workaround.)

This change sets `spark.executor.cores` to `10`, if it is unset. This allows
integration tests to work around the failure seen in `parquet_test.py:test_small_file_memory`,
where the `COALESCING` Parquet reader's thread pool accidentally uses 128 threads with 8MB memory
each, thus consuming the entire heap.

Note that this is a bit of a workaround.  A more robust solution would be to scale the Parquet
reader's buffers based on the amount of available memory, and the number of threads.

Signed-off-by: MithunR <mythrocks@gmail.com>