Support Retry for GpuTopN and GpuSortEachBatchIterator (#9164)
For GpuSortEachBatchIterator, this change introduces a new API fullySortBatchAndCloseWithRetry in GpuSorter, which replaces the call to the original fullySortBatch in GpuSortEachBatchIterator.

GpuTopN is a little complicated. It is not a good idea to leverage this new API fullySortBatchAndCloseWithRetry, otherwise we will get a nested retry block. Instead we only need to retry the whole processing that returns a pending result for the next iteration. It also supports Split and Retry. Besides, retry is also added to the processing of applying the offset after getting the final result.
---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>