Integration tests: Install specific fastparquet version.
Fixes https://github.com/NVIDIA/spark-rapids/issues/9603.

This commit changes the integration test setup to specifically install
fastparquet-0.8.3.

Prior to this change, when the fastparquet version is not specified, the pip
install caused 0.5.0 to be installed on some nodes, e.g. on Dataproc 2.0
(with Spark 3.1.1).
The older fastparquet versions do not support reading the contents of input
directories recursively, causing the tests to fail.

Note that this change doesn't bump the version all the way to 2023.8.0, so as
to preserve compatibility with Dataproc 2.0.  v0.8.3 seems to have the broadest
support.

Signed-off-by: MithunR <mythrocks@gmail.com>