Running window optimization for LAST()
Fixes #9520.
Fixes #9299.

This is a followup to the changes in #9489, which adds the running
window optimization to the `FIRST()` window fuction.

This commit adds running window support for the `LAST()` window
function, with support for both `ROWS` and `RANGE` based window
specifications.
This change should allow the `LAST()` aggregation to run in multiple
batches if the window is `[UNBOUNDED PRECEDING, CURRENT ROW`]. This
will allow for much larger window and group sizes, because the group
no longer needs to fit in GPU memory. This should help mitigate
out-of-memory errors with `LAST()` window aggregations.

Signed-off-by: MithunR <mythrocks@gmail.com>