Use the original output when creating a canonicalized GPU file scan.
The original output should be used when creating a canonicalized GPU file scan, because
the rule `prunePartitionForFileSourceScan` in Plugin may remove the partition columns
that are not used by the first downstream ProjectExec for some patterns, leading to some
partition columns not exist in the finalized output.

Then the `AttributeReference`s in some filters but not in the output will not be canonicalized.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>