Fix ANSI mode failures in subquery_test.py.
Fixes #11029.

Some tests in subquery_test.py fail when run with ANSI mode enabled,
because certain array columns are accessed with invalid indices. These
tests predate the availability of ANSI mode in Spark.

This commit modifies the tests so that the generated data is
appropriately sized for the query.

There is no loss of test coverage; failure cases for invalid index
values in array columns are already tested as part of
`array_test::test_array_item_ansi_fail_invalid_index`

Signed-off-by: MithunR <mithunr@nvidia.com>