[Spark 4.0] Address test failures in cast_test.py [databricks] (#11559)
* Spark 4:  Addressed cast_test.py failures.

Fixes #11009 and #11530.

This commit addresses the test failures in cast_test.py, on Spark 4.0.
These generally have to do with changes in behaviour of Spark when
ANSI mode is enabled.  In these cases, the tests have been split out into ANSI=on and ANSI=off.  

The bugs uncovered from the tests have been spun into their own issues;  fixing all of them was beyond the scope of this change.

Signed-off-by: MithunR <mithunr@nvidia.com>