Fix `collection_ops_tests` for Spark 4.0 [databricks] (#11414)
* Fix collection_ops_tests for Spark 4.0.

Fixes #11011.

This commit fixes the failures in `collection_ops_tests` on Spark 4.0.

On all versions of Spark, when a Sequence is collected with rows that exceed MAX_INT,
an exception is thrown indicating that the collected Sequence/array is
larger than permissible. The different versions of Spark vary in the
contents of the exception message.

On Spark 4, one sees that the error message now contains more
information than all prior versions, including:
1. The name of the op causing the error
2. The errant sequence size

This commit introduces a shim to make this new information available in
the exception.

Note that this shim does not fit cleanly in RapidsErrorUtils, because
there are differences within major Spark versions. For instance, Spark
3.4.0-1 have a different message as compared to 3.4.2 and 3.4.3.
Likewise, the differences in 3.5.0, 3.5.1, 3.5.2.

Signed-off-by: MithunR <mithunr@nvidia.com>

* Fixed formatting error.

* Review comments.

This moves the construction of the long-sequence error strings into
RapidsErrorUtils.  The process involved introducing many new RapidsErrorUtils
classes, and using mix-ins of concrete implementations for the error-string
construction.

* Added missing shim tag for 3.5.2.

* Review comments: Fixed code style.

* Reformatting, per project guideline.

* Fixed missed whitespace problem.

---------

Signed-off-by: MithunR <mithunr@nvidia.com>