Fix string_test for [databricks] 14.3
Fixes #11535.

This commit fixes the failure of `test_startswith`
and `test_endswith` in `string_test.py` on Databricks 14.3.

This is, in effect, a follow-on to the changes introduced in #11247,
where `test_endswith` was skipped for Databricks 4.0, on account of
https://issues.apache.org/jira/browse/SPARK-48995.  It appears that
that bug afflicts Databricks 14.3 as well.

This commit handles both `test_startswith` and `test_endswith` for
Databricks 14.3 and Apache Spark 4.0.

Signed-off-by: MithunR <mithunr@nvidia.com>