Fix Parquet Writer tests on Databricks 14.3
Fixes #11534.

This commit fixes the test failures in `parquet_write_test.py`, as
listed on #11534.

This is an extension of the changes made in #11615, which attempted to
address similar failures on Apache Spark 4.

Most of the test failures pertain to legacy Parquet writer settings and
conf keys which were removed on Spark 4.  A stray test involves a change
in how the test gleans description strings from exceptions.

Signed-off-by: MithunR <mithunr@nvidia.com>