Fix skipping fixed_length_char ORC tests on [databricks] > 13.3 (#11652)
Fixes #11528.

This commit addresses the failure of
`orc_test.py::test_project_fallback_when_reading_hive_fixed_length_char`
when run on Databricks 13.3+.

This test was skipped with an incorrect reason statement in #9646, to
address a failure on Databricks 13.3.  That failure was not the result
of a data mismatch.  It failed because the ProjectExec didn't fall back
to CPU.  (In fact, with Databricks 13.3 onwards, the ProjectExec is not
involved in `SELECT *` queries.)

As an aside, the same test is now skipped on Databricks 13.3 and 14.3,
because the error condition it tests does not fire on those versions.

Signed-off-by: MithunR <mithunr@nvidia.com>