Fix aqe_test failures on [databricks] 14.3.
Fixes #11643.

This commit fixes the AQE/DPP tests that were reported in #11643 to
be failing on Databricks 14.3.

This is the result of a deficient shim for GpuSubqueryBroadcastMeta
being active for Databricks 14.3.  The deficient shim errantly
extended the non-Databricks base shim.

This commit moves the commonality in Databricks shims to a common
base class that is then customized for the changes in Databricks 14.3.

Signed-off-by: MithunR <mithunr@nvidia.com>