Make serialization key checks uniform for Hive Text file read/write.
This is a minor change, to bring the serialization format checks on the Read and Write
sides of Hive text files in line with each other.

The read side would check if the serialization-key is set to empty (""):
https://github.com/NVIDIA/spark-rapids/blob/aa2da410511d8a737e207257769ec662a79174fe/sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/HiveProviderImpl.scala#L155-L161

The write side was checking for the expected default: "1":
https://github.com/NVIDIA/spark-rapids/blob/aa2da410511d8a737e207257769ec662a79174fe/sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/GpuHiveFileFormat.scala#L130-L136

This caused confusion in behaviour, as voiced in https://github.com/NVIDIA/spark-rapids/issues/11803.

The change in this commit makes the checks uniformly conservative.

Signed-off-by: MithunR <mithunr@nvidia.com>